refactor: separate configuration module
