+++ src/main/kotlin/com/kroffle/knitting/infra/properties/SelfProperties.kt
@@ -1,4 +1,4 @@
-package com.kroffle.knitting.infra.configuration
+package com.kroffle.knitting.infra.properties
 
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
